requirements.txt
aenum==3.1.11
aiofile==3.8.1
aiohttp==3.8.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.26.2
azure-identity==1.12.0
azure-quantum==0.27.244707
azure-storage-blob==12.14.1
biopython==1.80
caio==0.9.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
fastapi==0.89.1
frozenlist==1.3.3
gremlinpython==3.6.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mpmath==1.2.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
nest-asyncio==1.5.6
ntlm-auth==1.5.0
numpy==1.24.1
oauthlib==3.2.2
pandas==1.5.3
pbr==5.11.1
ply==3.11
portalocker==2.7.0
protobuf==3.20.3
psutil==5.9.4
py==1.11.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pymongo==4.3.3
pyqir-generator==0.4.2a1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-markdown-math==0.8
pytz==2022.7.1
pywin32==305
qiskit==0.40.0
qiskit-aer==0.11.2
qiskit-ibmq-provider==0.19.2
qiskit-ionq==0.3.10
qiskit-qir==0.2.0
qiskit-terra==0.23.0
requests==2.28.2
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
retry==0.9.2
rustworkx==0.12.1
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
stevedore==4.1.1
sympy==1.11.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
websocket-client==1.5.0
websockets==10.4
wrapt==1.14.1
yarl==1.8.2
zipp==3.12.0
